hotfix: snapshot create optimized
